Tidy up types, rename shared-types worker-interface-types
