+++ packages/common/src/index.ts
@@ -2,4 +2,4 @@ export * from './types';
 export * from './utils';
 export * from './constants';
 export * from './languages';
-export * from './shared-types';
+export * from './worker-interface-types';
