+++ worker-interface/scala/.gitignore
@@ -1,0 +1,3 @@
+target/
+project/target/
+project/project/
